+++ client/postcss.config.js
@@ -1,0 +1,15 @@
+import autoprefixer from 'autoprefixer';
+import cssnano from 'cssnano';
+
+export default {
+    plugins: [
+        autoprefixer({
+            cascade: false,
+            supports: false
+        }),
+
+        cssnano({
+            preset: 'default',
+        })
+    ]
+};